🧩 GeminiChat Component:
Modularization
We've successfully modularized the GeminiChat component, enhancing its scalability and maintainability.

⚡ Default Integration of Gemini 2.5 Flash:

The component now utilizes the experimental Gemini 2.5 Flash model by default, offering improved performance and advanced reasoning capabilities.
